+++ lib/router.js
@@ -2794,8 +2794,8 @@ router.get('/jijitang/publication', lazyloadRouteHandler('./routes/jijitang/publ
 // router.get('/xwlb', lazyloadRouteHandler('./routes/xwlb/index'));
 
 // 端传媒
-router.get('/initium/:type?/:language?', lazyloadRouteHandler('./routes/initium/full'));
-router.get('/theinitium/:model/:type?/:language?', lazyloadRouteHandler('./routes/initium/full'));
+// router.get('/initium/:type?/:language?', lazyloadRouteHandler('./routes/initium/full'));
+// router.get('/theinitium/:model/:type?/:language?', lazyloadRouteHandler('./routes/initium/full'));
 
 // Grub Street
 router.get('/grubstreet', lazyloadRouteHandler('./routes/grubstreet/index'));
