+++ lib/v2/theinitium/maintainer.js
@@ -1,0 +1,6 @@
+module.exports = {
+    '/author/:type/:language?': ['AgFlore'],
+    '/channel/:type?/:language?': ['prnake'],
+    '/follow/articles/:language?': ['AgFlore'],
+    '/tags/:type/:language?': ['AgFlore'],
+};
