+++ lib/v2/theinitium/router.js
@@ -1,0 +1,3 @@
+module.exports = (router) => {
+    router.get('/:model?/:type?/:language?', require('./full'));
+};
